added Python message class
